Update and rename Script1(working model)1.py to Final_Project_1.py
